refactor: ScheduleBox in project page
