+++ components/ScheduleBox/Content/Spacer.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const Spacer = styled.div`
+  flex: 1;
+  flex-shrink: 0;
+`;
+
+export default Spacer;
