+++ components/ScheduleBox/Content/index.tsx
@@ -1,0 +1,40 @@
+import styled from 'styled-components';
+import { FC, ReactNode } from 'react';
+import { media } from '../../../styles/theme';
+import StepLottie, { LottieProps } from '../../StepLottie';
+import Description from './Description';
+import Animation from './Animation';
+import Spacer from './Spacer';
+
+interface ScheduleBoxContentProps {
+  desc: ReactNode;
+  lottie: LottieProps;
+  className?: string;
+}
+
+const Content:FC<ScheduleBoxContentProps> = ({ className, desc, lottie }) => (
+  <div className={className}>
+    <Description>{desc}</Description>
+    <Spacer className="mobile__none" />
+    <Animation>
+      <div className="lottie-container">
+        <StepLottie lottie={lottie} />
+      </div>
+    </Animation>
+  </div>
+);
+
+export default styled(Content)`
+  display: flex;
+  ${media.mobile} {
+    justify-content: space-between;
+    align-items: center;
+  }
+  .lottie-container {
+    position: relative;
+    height: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: flex-end;
+  }
+`;
